demo: add disabled demo
