+++ components/tooltip/demo/disabled-children.md
@@ -1,0 +1,7 @@
+## zh-CN
+
+Disabled 子元素。
+
+## en-US
+
+Disabled wrapper.
