+++ components/tooltip/demo/disabled.tsx
@@ -1,28 +1,17 @@
-import React from 'react';
-import { Button, Checkbox, Input, Select, Space, Tooltip, InputNumber } from 'antd';
+import React, { useState } from 'react';
+import { Space, Button, Tooltip } from 'antd';
 
-const WrapperTooltip: React.FC<{ children?: React.ReactNode }> = (props) => (
-  <Tooltip title="Thanks for using antd. Have a nice day!" {...props} />
-);
+const App: React.FC = () => {
+  const [disabled, setDisabled] = useState(true);
 
-const App: React.FC = () => (
-  <Space>
-    <WrapperTooltip>
-      <Button disabled>Disabled</Button>
-    </WrapperTooltip>
-    <WrapperTooltip>
-      <Input disabled placeholder="disabled" />
-    </WrapperTooltip>
-    <WrapperTooltip>
-      <InputNumber disabled />
-    </WrapperTooltip>
-    <WrapperTooltip>
-      <Checkbox disabled />
-    </WrapperTooltip>
-    <WrapperTooltip>
-      <Select disabled />
-    </WrapperTooltip>
-  </Space>
-);
+  return (
+    <Space>
+      <Button onClick={() => setDisabled(!disabled)}>{disabled ? 'Enable' : 'Disable'}</Button>
+      <Tooltip title={disabled ? '' : 'prompt text'}>
+        <span>Tooltip will show on mouse enter.</span>
+      </Tooltip>
+    </Space>
+  );
+};
 
 export default App;
