+++ components/tooltip/index.zh-CN.md
@@ -29,7 +29,8 @@ demo:
 <code src="./demo/colorful.tsx">多彩文字提示</code>
 <code src="./demo/render-panel.tsx" debug>_InternalPanelDoNotUseOrYouWillBeFired</code>
 <code src="./demo/debug.tsx" debug>Debug</code>
-<code src="./demo/disabled.tsx" debug>禁用</code>
+<code src="./demo/disabled.tsx">禁用</code>
+<code src="./demo/disabled-children.tsx" debug>禁用子元素</code>
 
 ## API
